moved visuals to own thing
